Fix: _train_one_helper bug fixed. global seed bug fixed
