+++ financial_loss_functions/config/hparams.json
@@ -8,22 +8,27 @@
         "BaseLSTM": {
             "model": {
                 "hidden_size": 16,
-                "num_layers": 8,
+                "num_layers": 4,
                 "dropout": 0.2,
                 "equal_prior": false
             },
             "optimizer": {
                 "lr": 1e-4,
-                "weight_decay": 1e-5
+                "weight_decay": 1e-2
             },
             "train" : {
-                "train_batch_size": 128,
-                "val_batch_size": 128,
+                "train_batch_size": 64,
+                "val_batch_size": 64,
                 "clip_grad_norm": 0.5,
                 "epochs": 200,
                 "early_stopping": true,
                 "early_stop_patience": 20,
                 "early_stop_min_delta": 1e-5
+            },
+            "scheduler": {
+                "factor": 0.5,
+                "patience": 10,
+                "min_lr": 1e-6
             }
         },
         "AttentionLSTM": {
@@ -70,7 +75,7 @@
                 "train_batch_size": 64,
                 "val_batch_size": 64,
                 "clip_grad_norm": 0.1,
-                "epochs": 500,
+                "epochs": 200,
                 "early_stopping": true,
                 "early_stop_patience": 20,
                 "early_stop_min_delta": 1e-5
@@ -98,7 +103,7 @@
                 "train_batch_size": 64,
                 "val_batch_size": 64,
                 "clip_grad_norm": 0.5,
-                "epochs": 500,
+                "epochs": 200,
                 "early_stopping": true,
                 "early_stop_patience": 20,
                 "early_stop_min_delta": 1e-5
