+++ financial_loss_functions/src/training/pipeline.py
@@ -5,9 +5,9 @@ import pandas as pd
 from torch import optim
 from pathlib import Path
 from src.utils.io import create_directory
-from src.utils.device import get_best_device
 from src.evaluation.evaluator import Evaluator
 from src.data_processing.loading import load_csv_files
+from src.utils.device import get_best_device, set_seed
 from src.data_processing.dataset import (
     Reshaper,
     calc_in_out_idx,
@@ -35,36 +35,8 @@ from src.models.registry import NNModelLibrary, TradModelLibrary
 # Add Best model ranker
 # Unit test NCO
 
-import random
-
-def set_seed(seed=50):
-    # 1. Basic Python and Numpy seeds
-    random.seed(seed)
-    np.random.seed(seed)
-    # os.environ['PYTHONHASHSEED'] = str(seed)
-    
-    # 2. Basic PyTorch seed (covers CPU)
-    torch.manual_seed(seed)
-    
-    # 3. NVIDIA CUDA Specifics
-    if torch.cuda.is_available():
-        torch.cuda.manual_seed(seed)
-        torch.cuda.manual_seed_all(seed) # for multi-GPU
-        # These two ensure deterministic behavior but may slow down training slightly
-        torch.backends.cudnn.deterministic = True
-        torch.backends.cudnn.benchmark = False
-        
-    # 4. Apple Silicon (MPS) Specifics
-    if hasattr(torch, "mps") and torch.backends.mps.is_available():
-        torch.mps.manual_seed(seed)
-        # Note: MPS is still maturing; some operations might not be 100% deterministic yet
-        
-    print(f"Seeds set to {seed} across all available backends.")
-
-set_seed(50)
-
-
 def _common_setup(paths_config):
+    set_seed(50) # Global seed for reproducibility
     # Create plots directory if it doesnt exist
     plots_dir = (Path(paths_config['artifacts']['plots']))
     create_directory(plots_dir)
